// frontend/src/api/models/JobType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overmind Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Valid job types for creation.
 * @export
 */
export const JobType = {
    AgentDiscovery: 'agent_discovery',
    JudgeScoring: 'judge_scoring',
    PromptTuning: 'prompt_tuning',
    ModelBacktesting: 'model_backtesting'
} as const;
export type JobType = typeof JobType[keyof typeof JobType];


export function instanceOfJobType(value: any): boolean {
    for (const key in JobType) {
        if (Object.prototype.hasOwnProperty.call(JobType, key)) {
            if (JobType[key as keyof typeof JobType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function JobTypeFromJSON(json: any): JobType {
    return JobTypeFromJSONTyped(json, false);
}

export function JobTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): JobType {
    return json as JobType;
}

export function JobTypeToJSON(value?: JobType | null): any {
    return value as any;
}

export function JobTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): JobType {
    return value as JobType;
}